Stability Test Suite (Not ready to review yet)
[problem]
Bugs reports coming in as crashes that have happened just once are hard
to reproduce.

[cause]
Some odd bugs only show up under load.

[solution]

A set of tests that do things to flush out odd bugs and leaks that
otherwise only show up in the wild and come in as nasty bug reports.
The intention is for these tests to be run regularly in an automated
fashion to reveal problems as soon as they are introduced.
Running under a tool such as Valgrind is desirable for some automated
runs to get the leak and invalid memory use reporting.
These tests should be built and run both in debug mode to pump the
assertions in the codebase and in release mode.

Error states detected are:
* Crashes of the test apps.
* Leaks reported by valgrind-like tools.
* Memory corruption reported by internal tools and Valgrind.
* Errors reported by the apps on stderr.
* Errors appearing in Dali log.

E.g.:

* Construct an Application object and destroy it in a loop.
* Construct and throw away an Adapter in a loop.
* Construct and throw away a PlatformAbstraction in a loop.
* Construct a stage and throw away ... rinse ...
* Construct lots of actors with different lifetime policies,
  adding and removing them in multiple places.

Change-Id: I22207197ec209b2e620d35592cd4aae3894f4425
Signed-off-by: Andrew Cox <andrew.cox@partner.samsung.com>