Fix for text-layout-example grab control not showing
The problem is caused by resize policy of the grab control being "natural size" which for
the case of control is the natural size of the background. In master branch the background is an
ImageActor so the natural size is the size of the image, but that's not the case anymore in new_mesh
branch, where the background is just a mesh.

Maybe actors should calculate its natural size based on the size of the geometry of its renderer, currently
actor return Vector3::ZERO as its natural size.

Change-Id: I96e49830b725a7d08d94737d06c7700292a61291