(TextLabel Example) Fixed Pan gesture handling
Problem:   When we are resizing and go past the minimum size of the container and then release, we have to move
           the distance we moved from the minimum size to the our release point again in order to make the
           container resize again.
Reason:    This was because we kept on changing mLayoutSize but did not reset it when the pan starts.
Solution:  Ensure we get the latest size of the container when the pan starts.

Change-Id: I3ded29691e8f1c405847f971d9a89d50aa448fdf