Changed DALI_*_DIR to DEMO_*_DIR
Scripts, examples and demo code should not use defines from other
parts of DALi. DALI_IMAGE_DIR is defined in Dali toolkit to specify
where the toolkit images are. This is used in the default toolkit
style file to specify the locations of the toolkit images.

It must not be overridden by applications.

Updated the demo style file to utilize a configured variable to point
at the installed demo style's images.

Change-Id: Ie43a6611a8db3fc0afc6a1cd4d9d6e48afef75e7
Signed-off-by: David Steele <david.steele@samsung.com>